+++ src/snowflake/cli/plugins/nativeapp/exceptions.py
@@ -14,7 +14,7 @@ class ApplicationPackageAlreadyExistsError(ClickException):
 
 
 class ApplicationPackageDoesNotExistError(ClickException):
-    """An application package of the specified name does not exist in the Snowflake acccount."""
+    """An application package of the specified name does not exist in the Snowflake account."""
 
     def __init__(self, name: str):
         super().__init__(