+++ RELEASE-NOTES.md
@@ -9,7 +9,8 @@
 * `snow snowpark package create`:
   * `--pypi-download` or `--yes` flags have no effect and will cause a warning.
     `create` command always checks against PyPi.
-  * `--allow-native-libraries` is renamed to `--allow-shared-libraries`. Old name of the flag is still supported.
+  * `--allow-native-libraries` is replaced with boolean flag `--allow-shared-libraries`.
+    Using old flag will cause a waring.
   * new `--ignore-anaconda` flag disables package lookup in Snowflake Anaconda channel.
     All dependencies will be downloaded from PyPi.
   * new `--skip-version-check` skips comparing versions of dependencies between requirements and Anaconda.
@@ -20,6 +21,7 @@
 * Added support for fully qualified image repository names in `spcs image-repository` commands.
 * Added `--if-not-exists` option to `create` commands for `service`, and `compute-pool`. Added `--replace` and `--if-not-exists` options for `image-repository create`.
 * Added support for python connector diagnostic report.
+* Added `snow app deploy` command that creates an application package and syncs the local changes to the stage without creating or updating the application.
 
 ## Fixes and improvements
 * Adding `--image-name` option for image name argument in `spcs image-repository list-tags` for consistency with other commands.
@@ -179,7 +181,7 @@
 # v1.2.0
 
 ## Backward incompatibility
-* Removed `snow streamlit create` command. Streamlit can be deployd using `snow streamlit deploy`
+* Removed `snow streamlit create` command. Streamlit can be deployed using `snow streamlit deploy`
 * Removed short option names in compute pool commands:
   * `-n` for `--name`, name of compute pool
   * `-d` for `--num`, number of pool's instances